upd model_objects
