+++ .gitmodules
@@ -2,4 +2,3 @@
     path = Ecdar-ProtoBuf
     url = git@github.com:Ecdar/Ecdar-ProtoBuf.git
     branch = SW5
-#	url = https://github.com/ECDAR-AAU-SW-P5/Ecdar-ProtoBuf.git
