+++ src/model_objects/declarations.rs
@@ -1,0 +1,57 @@
+use crate::model_objects::Component;
+use edbm::util::constraints::ClockIndex;
+use serde::{Deserialize, Serialize};
+use std::collections::{HashMap, HashSet};
+
+pub trait DeclarationProvider {
+    fn get_declarations(&self) -> &Declarations;
+}
+
+impl DeclarationProvider for Component {
+    fn get_declarations(&self) -> &Declarations {
+        &self.declarations
+    }
+}
+
+/// The declaration struct is used to hold the indices for each clock, and is meant to be the owner of int variables once implemented
+#[derive(Debug, Deserialize, Clone, PartialEq, Eq, Serialize)]
+pub struct Declarations {
+    pub ints: HashMap<String, i32>,
+    pub clocks: HashMap<String, ClockIndex>,
+}
+
+impl Declarations {
+    pub fn empty() -> Declarations {
+        Declarations {
+            ints: HashMap::new(),
+            clocks: HashMap::new(),
+        }
+    }
+
+    pub fn remove_clock(&mut self, clock: &str) {
+        self.clocks.remove(clock);
+    }
+
+    pub fn get_clock_count(&self) -> usize {
+        self.clocks.values().collect::<HashSet<_>>().len()
+    }
+
+    pub fn set_clock_indices(&mut self, start_index: ClockIndex) {
+        for (_, v) in self.clocks.iter_mut() {
+            *v += start_index
+        }
+    }
+
+    pub fn get_clock_index_by_name(&self, name: &str) -> Option<&ClockIndex> {
+        self.clocks.get(name)
+    }
+
+    /// Gets the name of a given `ClockIndex`.
+    /// Returns `None` if it does not exist in the declarations
+    pub fn get_clock_name_by_index(&self, index: ClockIndex) -> Option<&String> {
+        self.clocks
+            .iter()
+            .find(|(_, v)| **v == index)
+            .map(|(k, _)| k)
+    }
+}
