+++ src/model_objects/queries.rs
@@ -1,25 +1,24 @@
-use crate::data_reader::parse_queries;
-use crate::model_objects::expressions;
+use crate::data_reader::parse_queries::parse_to_expression_tree;
+use crate::model_objects::expressions::QueryExpression;
+use serde::de::Error;
 use serde::{Deserialize, Deserializer};
 
 /// The struct containing a single query
 #[derive(Debug, Deserialize, Clone)]
 pub struct Query {
     #[serde(deserialize_with = "decode_query")]
-    pub query: Option<expressions::QueryExpression>,
+    pub query: Option<QueryExpression>,
     pub comment: String,
 }
 
 impl Query {
-    pub fn get_query(&self) -> &Option<expressions::QueryExpression> {
+    pub fn get_query(&self) -> &Option<QueryExpression> {
         &self.query
     }
 }
 
 /// Function used for deserializing queries
-pub fn decode_query<'de, D>(
-    deserializer: D,
-) -> Result<Option<expressions::QueryExpression>, D::Error>
+pub fn decode_query<'de, D>(deserializer: D) -> Result<Option<QueryExpression>, D::Error>
 where
     D: Deserializer<'de>,
 {
@@ -28,12 +27,16 @@ where
         return Ok(None);
     }
 
-    let queries = parse_queries::parse_to_expression_tree(&s).unwrap();
-    if queries.len() > 1 {
-        panic!("Could not parse query {} contains multiple queries", s);
-    } else if queries.is_empty() {
-        panic!("Could not parse query {} contains no queries", s);
-    } else {
-        Ok(queries.into_iter().next())
+    let queries = parse_to_expression_tree(&s).unwrap();
+    match queries.len() {
+        0 => Err(Error::custom(format!(
+            "Could not parse query {} contains no queries",
+            s
+        ))),
+        1 => Ok(queries.into_iter().next()),
+        _ => Err(Error::custom(format!(
+            "Could not parse query {} contains multiple queries",
+            s
+        ))),
     }
 }
